explore: move exploration to xtc.search.explore
