+++ docs/tutorials/explore_optimizers.py
@@ -7,106 +7,26 @@ app = marimo.App(width="medium")
 @app.cell
 def _():
     import marimo as mo
-    import subprocess
-    from pathlib import Path
-    import numpy as np
-    import random
-    import xtc.cli.explore as explore
-    OPERATORS = explore.OPERATORS
-    STRATEGIES_ALIASES = explore.STRATEGIES_ALIASES
-    return OPERATORS, STRATEGIES_ALIASES, explore, mo, np, random
+    import xtc.search.explore as explore
+    return explore, mo
 
 
 @app.cell
-def _(OPERATORS, explore, np, random):
-    import multiprocessing
-
-    class Args:
-        def __init__(self, **overrides):
-            self.operator = "matmul"
-            self.op_name = None
-            self.ops_list = False
-            self.func_name = None
-            self.strategy = None
-            self.search = "random"
-            self.backends = ["mlir"]
-            self.optimizer = "random-forest-default"
-            self.data = None
-            self.dims = None
-            self.huge_pages = True
-            self.test = []
-            self.opt_level = 4
-            self.dtype = None
-            self.trials = 100
-            self.threads = 1
-            self.max_unroll = 512
-            self.seed = 0
-            self.output = "results.csv"
-            self.resume = False
-            self.append = False
-            self.eval = "eval"
-            self.repeat = 1
-            self.number = 1
-            self.min_repeat_ms = 100
-            self.validate = None
-            self.save_temps = None
-            self.save_temps_dir = "./save_temps_dir"
-            self.explore_dir = "."
-            self.optimizer_config = None
-            self.child = True
-            self.bare_ptr = False
-            self.jobs = max(1, multiprocessing.cpu_count() // 2)
-            self.execute = True
-            self.peak_flops = None
-            self.mlir_prefix = None
-            self.batch = 1
-            self.debug = None
-            self.debug_compile = None
-            self.debug_xtc = None
-            self.debug_optimizer = None
-            self.quiet = None
-            self.dump = None
-
-            self.__dict__.update(overrides)
-
-            if not self.func_name:
-                self.func_name = self.operator
-
-            if not self.strategy:
-                self.strategy = OPERATORS[self.operator]["default_strategy"]
-
-            if not self.dims:
-                self.dims = OPERATORS[self.operator]["default_dims"]
-
-            if not self.dtype:
-                self.dtype = OPERATORS[self.operator]["default_type"]
-
-            if self.op_name:
-                self.dims = explore.get_operation_dims(self.operator, self.op_name)
-
-            # backend validation
-            for backend in self.backends:
-                assert backend in OPERATORS[self.operator]["backends"], (
-                    f"backend {backend} not available for operator {self.operator}"
-                )
-
-            if self.seed >= 0:
-                np.random.seed(self.seed)
-                random.seed(self.seed)
-
-            explore.setup_args(self)
+def _(explore):
+    def Args(**overrides):
+        return explore.default_exploration_config(**overrides)
     return (Args,)
 
 
 @app.cell
-def _(STRATEGIES_ALIASES, mo):
-    from xtc.artifacts import list_operations 
+def _(mo):
+    from xtc.artifacts import list_operations
     from xtc.search.strategies import Strategies
     from xtc.search.optimizers import Optimizers
 
     operators = ["matmul","conv2d"]
     backends = ["mlir","tvm","jir"]
-    strategies = list(STRATEGIES_ALIASES.keys()) + list(Strategies.names()) 
+    strategies = list(Strategies.names())
     backend_ui = mo.ui.multiselect(options=backends,value=["mlir"],label="backends:")
     operator_ui = mo.ui.radio(options=operators,value="matmul",label="operator:")
     def op_names(operator):
@@ -118,15 +38,20 @@ def _(STRATEGIES_ALIASES, mo):
         )
     strategy_ui = mo.ui.dropdown(
         options=strategies,
-        value=strategies[-1],
+        value="tile_oo",
         label="strategy:"
     )
+    strategy_param_ui = mo.ui.text(
+        value="",
+        label="prt strategy scheme:",
+        placeholder="e.g. PPWRPRP"
+    )
     seed_ui = mo.ui.number(start=0,stop=200,label="seed:")
-    trials_ui = mo.ui.number(start=0,stop=5000,value=100,label="trials:")
+    trials_ui = mo.ui.number(start=0,stop=2048,value=64,label="trials:")
 
     search_types = ["random","iterative","exhaustive"]
     search_select = mo.ui.radio(search_types,value=search_types[0],label="search:")
-    batch_ui = mo.ui.number(start=1,stop=50,value=5,label="batch size:")
+    batch_ui = mo.ui.number(start=1,stop=64,value=8,label="batch size:")
 
 
 
@@ -167,6 +92,7 @@ def _(STRATEGIES_ALIASES, mo):
         search_select,
         seed_ui,
         strategy_ui,
+        strategy_param_ui,
         trials_ui,
     )
 
@@ -193,6 +119,7 @@ def _(
     search_select,
     seed_ui,
     strategy_ui,
+    strategy_param_ui,
     trials_ui,
 ):
     search_ui = [search_select]
@@ -202,7 +129,10 @@ def _(
         optimizer_config_ui = get_optimizer_config_ui(optimizer_ui.value, batch_ui.value)
         search_ui += [mo.vstack([batch_ui,optimizer_ui]), optimizer_config_ui]
 
-    vstack_elements = [backend_ui,operator_ui, op_names(operator_ui.value),strategy_ui,trials_ui,seed_ui]
+    vstack_elements = [backend_ui, operator_ui, op_names(operator_ui.value), strategy_ui]
+    if strategy_ui.value == "prt":
+        vstack_elements.append(strategy_param_ui)
+    vstack_elements += [trials_ui, seed_ui]
     mo.vstack(vstack_elements, justify="start")
     return optimizer_config_ui, search_ui
 
@@ -233,6 +163,7 @@ def _(
     search_select,
     seed_ui,
     strategy_ui,
+    strategy_param_ui,
     trials_ui,
 ):
     import tempfile
@@ -253,10 +184,15 @@ def _(
         pass
 
     def set_explore_args(config_path):
+        strategy_param = strategy_param_ui.value if strategy_ui.value == "prt" else ""
+        if strategy_param:
+            strategy = f"{strategy_ui.value}:{strategy_param}"
+        else:
+            strategy = strategy_ui.value
         args = Args(
             operator = operator_ui.value,
             backends = backend_ui.value,
-            strategy = strategy_ui.value,
+            strategy = strategy,
             search = search_select.value,
             seed = seed_ui.value,
             batch = batch_ui.value,
@@ -278,7 +214,8 @@ def _(
             args.output = output_path
 
             #mo.output.append("optimizing...")
-            explore.optimize(args)
+            exploration = explore.Exploration(args)
+            exploration()
 
             best = 0
             best_info = ""
@@ -297,9 +234,8 @@ def _(
 
     def run_loop_explore():
         best, best_info = loop_explore()
-        #mo.output.append(f"best peak perf was {best}")
-        return f"best peak perf was {best}\n\n{best_info}"
-        #return mo.md(f"best peak perf was {best}")
+        return f"Best peak found is {best*100:.2f}% for {best_info}"
+
     return (run_loop_explore,)
