+++ src/xtc/artifacts/__init__.py
@@ -2,7 +2,6 @@
 # SPDX-License-Identifier: BSD-3-Clause
 # Copyright (c) 2024-2026 The XTC Project Authors
 #
-# type: ignore
 from .operations import (
     register_operation,
     get_operation,
@@ -17,3 +16,10 @@ from .register_subgraph_operations import (
 
 _register_operations()
 _register_subgraph_operations()
+
+__all__ = [
+    "register_operation",
+    "get_operation",
+    "list_operations",
+    "has_operation",
+]
