+++ src/xtc/utils/numpy.py
@@ -2,14 +2,14 @@
 # SPDX-License-Identifier: BSD-3-Clause
 # Copyright (c) 2024-2026 The XTC Project Authors
 #
-from typing import Any
+from typing import Any, Sequence
 import numpy as np
 import numpy.typing
 
 from .math import mulall
 
 
-def np_init(shape: tuple, dtype: str) -> numpy.typing.NDArray[Any]:
+def np_init(shape: Sequence[int], dtype: str) -> numpy.typing.NDArray[Any]:
     """
     Initialize and return a NP array filled
     with numbers in [1, 9].
