+++ tests/pytest/unit/cli/test_explore_csv_callback.py
@@ -1,7 +1,7 @@
 import csv
 from pathlib import Path
 
-from xtc.cli.explore import CSVCallback
+from xtc.search.explore import CSVCallback
 
 
 def _read_rows(path: Path) -> list[list[str]]:
@@ -17,7 +17,7 @@ def test_csv_callback_resume_dedup_skips_existing_and_keeps_new(tmp_path: Path):
     with output.open("w", newline="") as out:
         writer = csv.writer(out, delimiter=",")
         writer.writerow(sample_names + ["X", "time", "peak", "backend"])
-        writer.writerow([8, 16, "[8; 16]", 0.2, 5.0, "mlir"])
+        writer.writerow([8, 16, "[8; 16]", 2.0, 0.5, "mlir"])
 
     cb = CSVCallback(
         str(output),
@@ -27,11 +27,11 @@ def test_csv_callback_resume_dedup_skips_existing_and_keeps_new(tmp_path: Path):
     )
 
     # Duplicate for same backend must be skipped in resume mode.
-    cb(([8, 16], 0, 0.2, "mlir"))
+    cb(([8, 16], 0, 3.0, "mlir"))
     # Different backend for same sample must be recorded.
-    cb(([8, 16], 0, 0.25, "tvm"))
+    cb(([8, 16], 0, 2.5, "tvm"))
     # Different sample must be recorded.
-    cb(([8, 32], 0, 0.3, "mlir"))
+    cb(([8, 32], 0, 3.0, "mlir"))
     del cb
 
     rows = _read_rows(output)
@@ -41,8 +41,10 @@ def test_csv_callback_resume_dedup_skips_existing_and_keeps_new(tmp_path: Path):
     assert rows[0] == ["M", "N", "X", "time", "peak", "backend"]
 
     # Check appended rows (order matters).
+    assert rows[1][0:2] == ["8", "16"]
+    assert rows[1][-3:] == ["2.0", "0.5", "mlir"]
     assert rows[2][0:2] == ["8", "16"]
-    assert rows[2][-1] == "tvm"
+    assert rows[2][-3:] == ["2.5", "0.4", "tvm"]
     assert rows[3][0:2] == ["8", "32"]
     assert rows[3][-1] == "mlir"
 
@@ -54,7 +56,7 @@ def test_csv_callback_default_mode_overwrites_existing_file(tmp_path: Path):
     with output.open("w", newline="") as out:
         writer = csv.writer(out, delimiter=",")
         writer.writerow(sample_names + ["X", "time", "peak", "backend"])
-        writer.writerow([9, 9, "[9; 9]", 0.9, 1.1, "mlir"])
+        writer.writerow([9, 9, "[9; 9]", 8.0, 0.25, "mlir"])
 
     cb = CSVCallback(
         str(output),
@@ -63,7 +65,7 @@ def test_csv_callback_default_mode_overwrites_existing_file(tmp_path: Path):
     )
 
     # Default mode is neither --resume nor --append, so file is rewritten.
-    cb(([2, 3], 0, 0.5, "mlir"))
+    cb(([2, 3], 0, 6.0, "mlir"))
     del cb
 
     rows = _read_rows(output)
@@ -80,7 +82,7 @@ def test_csv_callback_append_mode_allows_duplicates(tmp_path: Path):
     with output.open("w", newline="") as out:
         writer = csv.writer(out, delimiter=",")
         writer.writerow(sample_names + ["X", "time", "peak", "backend"])
-        writer.writerow([4, 4, "[4; 4]", 0.1, 10.0, "mlir"])
+        writer.writerow([4, 4, "[4; 4]", 10.0, 0.1, "mlir"])
 
     cb = CSVCallback(
         str(output),
@@ -90,7 +92,7 @@ def test_csv_callback_append_mode_allows_duplicates(tmp_path: Path):
     )
 
     # Same sample/backend is allowed in append mode.
-    cb(([4, 4], 0, 0.12, "mlir"))
+    cb(([4, 4], 0, 1.6, "mlir"))
     del cb
 
     rows = _read_rows(output)
@@ -112,9 +114,9 @@ def test_csv_callback_append_writes_header_for_new_file(tmp_path: Path):
         sample_names=["M", "N"],
         append=True,
     )
-    cb(([2, 3], 0, 0.5, "mlir"))
+    cb(([2, 3], 0, 8.0, "mlir"))
     del cb
 
     rows = _read_rows(output)
     assert rows[0] == ["M", "N", "X", "time", "peak", "backend"]
-    assert rows[1][0:2] == ["2", "3"]
+    assert rows[1] == ["2", "3", "[2; 3]", "8.0", "0.25", "mlir"]
